+++ trace/src/render/render_graph/ToneMapPass.cpp
@@ -114,8 +114,8 @@ namespace trace {
 				RenderGraphFrameData* graph_data = renderer->GetRenderGraphData(render_graph_index);
 				Viewport view_port = m_renderer->_viewPort;
 				Rect2D rect = m_renderer->_rect;
-				view_port.width = width;
-				view_port.height = height;
+				view_port.width = static_cast<float>(width);
+				view_port.height = static_cast<float>(height);
 
 				rect.right = width;
 				rect.bottom = height;
